+++ examples/chat/main.go
@@ -1,19 +1,20 @@
-// Simple Go chat client for https://github.com/centrifugal/centrifuge/tree/master/examples/events example.
 package main
 
 import (
 	"bufio"
 	"encoding/json"
 	"log"
+	"net/http"
 	"os"
+
+	_ "net/http/pprof"
 
 	"github.com/centrifugal/centrifuge-go"
 	"github.com/dgrijalva/jwt-go"
 )
 
-// Actually in real life clients should never know secret key.
-// This is only for example purposes to quickly generate JWT for
-// connection.
+// In real life clients should never know secret key. This is only for example
+// purposes to quickly generate JWT for connection.
 const exampleTokenHmacSecret = "secret"
 
 func connToken(user string, exp int64) string {
@@ -30,26 +31,47 @@ func connToken(user string, exp int64) string {
 	return t
 }
 
-// ChatMessage is chat app specific message struct.
+// ChatMessage is chat example specific message struct.
 type ChatMessage struct {
 	Input string `json:"input"`
 }
 
 type eventHandler struct{}
 
-func (h *eventHandler) OnConnect(c *centrifuge.Client, e centrifuge.ConnectEvent) {
+func (h *eventHandler) OnConnect(_ *centrifuge.Client, e centrifuge.ConnectEvent) {
 	log.Printf("Connected to chat with ID %s", e.ClientID)
-	return
 }
 
-func (h *eventHandler) OnError(c *centrifuge.Client, e centrifuge.ErrorEvent) {
+func (h *eventHandler) OnError(_ *centrifuge.Client, e centrifuge.ErrorEvent) {
 	log.Printf("Error: %s", e.Message)
-	return
 }
 
-func (h *eventHandler) OnDisconnect(c *centrifuge.Client, e centrifuge.DisconnectEvent) {
+func (h *eventHandler) OnMessage(_ *centrifuge.Client, e centrifuge.MessageEvent) {
+	log.Printf("Message from server: %s", string(e.Data))
+}
+
+func (h *eventHandler) OnDisconnect(_ *centrifuge.Client, e centrifuge.DisconnectEvent) {
 	log.Printf("Disconnected from chat: %s", e.Reason)
-	return
+}
+
+func (h *eventHandler) OnServerSubscribe(_ *centrifuge.Client, e centrifuge.ServerSubscribeEvent) {
+	log.Printf("Subscribe to server-side channel %s: (resubscribe: %t, recovered: %t)", e.Channel, e.Resubscribed, e.Recovered)
+}
+
+func (h *eventHandler) OnServerUnsubscribe(_ *centrifuge.Client, e centrifuge.ServerUnsubscribeEvent) {
+	log.Printf("Unsubscribe from server-side channel %s", e.Channel)
+}
+
+func (h *eventHandler) OnServerJoin(_ *centrifuge.Client, e centrifuge.ServerJoinEvent) {
+	log.Printf("Server-side join to channel %s: %s (%s)", e.Channel, e.User, e.Client)
+}
+
+func (h *eventHandler) OnServerLeave(_ *centrifuge.Client, e centrifuge.ServerLeaveEvent) {
+	log.Printf("Server-side leave from channel %s: %s (%s)", e.Channel, e.User, e.Client)
+}
+
+func (h *eventHandler) OnServerPublish(_ *centrifuge.Client, e centrifuge.ServerPublishEvent) {
+	log.Printf("Publication from server-side channel %s: %s", e.Channel, e.Data)
 }
 
 func (h *eventHandler) OnPublish(sub *centrifuge.Subscription, e centrifuge.PublishEvent) {
@@ -58,15 +80,15 @@ func (h *eventHandler) OnPublish(sub *centrifuge.Subscription, e centrifuge.Publ
 	if err != nil {
 		return
 	}
-	log.Printf("Someone says: %s", chatMessage.Input)
+	log.Printf("Someone says via channel %s: %s", sub.Channel(), chatMessage.Input)
 }
 
 func (h *eventHandler) OnJoin(sub *centrifuge.Subscription, e centrifuge.JoinEvent) {
-	log.Printf("Someone joined: user id %s, client id %s", e.User, e.Client)
+	log.Printf("Someone joined %s: user id %s, client id %s", sub.Channel(), e.User, e.Client)
 }
 
 func (h *eventHandler) OnLeave(sub *centrifuge.Subscription, e centrifuge.LeaveEvent) {
-	log.Printf("Someone left: user id %s, client id %s", e.User, e.Client)
+	log.Printf("Someone left %s: user id %s, client id %s", sub.Channel(), e.User, e.Client)
 }
 
 func (h *eventHandler) OnSubscribeSuccess(sub *centrifuge.Subscription, e centrifuge.SubscribeSuccessEvent) {
@@ -77,24 +99,40 @@ func (h *eventHandler) OnSubscribeError(sub *centrifuge.Subscription, e centrifu
 	log.Printf("Subscribed on channel %s failed, error: %s", sub.Channel(), e.Error)
 }
 
-func (h *eventHandler) OnUnsubscribe(sub *centrifuge.Subscription, e centrifuge.UnsubscribeEvent) {
+func (h *eventHandler) OnUnsubscribe(sub *centrifuge.Subscription, _ centrifuge.UnsubscribeEvent) {
 	log.Printf("Unsubscribed from channel %s", sub.Channel())
 }
 
-func main() {
-	url := "ws://localhost:8000/connection/websocket?format=protobuf"
+func newClient(handler *eventHandler) *centrifuge.Client {
+	wsURL := "ws://localhost:8000/connection/websocket"
+	c := centrifuge.New(wsURL, centrifuge.DefaultConfig())
 
-	log.Printf("Connect to %s\n", url)
-	log.Printf("Print something and press ENTER to send\n")
-
-	c := centrifuge.New(url, centrifuge.DefaultConfig())
-	// Uncomment to make it work with Centrifugo and JWT auth.
+	// Uncomment to make it work with Centrifugo and its JWT auth.
 	//c.SetToken(connToken("49", 0))
-	defer c.Close()
-	handler := &eventHandler{}
+
 	c.OnConnect(handler)
-	c.OnError(handler)
 	c.OnDisconnect(handler)
+	c.OnMessage(handler)
+	c.OnError(handler)
+
+	c.OnServerPublish(handler)
+	c.OnServerSubscribe(handler)
+	c.OnServerUnsubscribe(handler)
+	c.OnServerJoin(handler)
+	c.OnServerLeave(handler)
+
+	return c
+}
+
+func main() {
+	go func() {
+		log.Println(http.ListenAndServe(":5000", nil))
+	}()
+
+	handler := &eventHandler{}
+
+	c := newClient(handler)
+	defer func() { _ = c.Close() }()
 
 	sub, err := c.NewSubscription("chat:index")
 	if err != nil {
@@ -113,23 +151,35 @@ func main() {
 		log.Fatalln(err)
 	}
 
+	pubText := func(text string) error {
+		msg := &ChatMessage{
+			Input: text,
+		}
+		data, _ := json.Marshal(msg)
+		_, err := sub.Publish(data)
+		return err
+	}
+
 	err = c.Connect()
 	if err != nil {
 		log.Fatalln(err)
 	}
+
+	err = pubText("hello")
+	if err != nil {
+		log.Printf("Error publish: %s", err)
+	}
+
+	log.Printf("Print something and press ENTER to send\n")
 
 	// Read input from stdin.
 	go func(sub *centrifuge.Subscription) {
 		reader := bufio.NewReader(os.Stdin)
 		for {
 			text, _ := reader.ReadString('\n')
-			msg := &ChatMessage{
-				Input: text,
-			}
-			data, _ := json.Marshal(msg)
-			err := sub.Publish(data)
+			err = pubText(text)
 			if err != nil {
-				log.Printf("publish error: %v", err)
+				log.Printf("Error publish: %s", err)
 			}
 		}
 	}(sub)