+++ transport_websocket.go
@@ -123,6 +123,7 @@ func (t *websocketTransport) reader() {
 			t.disconnect = disconnect
 			return
 		}
+		//println("<----", strings.Trim(string(data), "\n"))
 	loop:
 		for {
 			decoder := newReplyDecoder(t.encoding, data)
@@ -159,6 +160,7 @@ func (t *websocketTransport) Write(cmd *protocol.Command, timeout time.Duration)
 	if timeout > 0 {
 		_ = t.conn.SetWriteDeadline(time.Now().Add(timeout))
 	}
+	//println("---->", strings.Trim(string(data), "\n"))
 	if t.encoding == protocol.TypeJSON {
 		err = t.conn.WriteMessage(websocket.TextMessage, data)
 	} else {